bloglist expansion, step9; blog can be deleted only by user who added blog

+++ controllers/blogs.js
@@ -49,8 +49,24 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id);
-  response.status(204).end();
+  console.log(request.token);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+  if (!request.token || !decodedToken.id) {
+    return response.status(401).json({ error: "token missing or invalid" });
+  }
+
+  const blog = await Blog.findById(request.params.id);
+  const user = await User.findById(decodedToken.id);
+
+  if (blog.user.toString() === user.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id);
+    response.status(204).end();
+  }
+
+  return response
+    .status(401)
+    .json({ error: "Wrong user trying to delete blog post" });
 });
 
 blogsRouter.put("/:id", async (request, response) => {
